# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/produtos": {
            "get": {
                "tags": [
                    "Produtos"
                ],
                "summary": "Exibe uma lista de produtos com filtros opcionais",
                "description": "Retorna uma lista de produtos com base nos filtros fornecidos (nome, descrição, preço, estoque)",
                "operationId": "getProdutos",
                "parameters": [
                    {
                        "name": "nome",
                        "in": "query",
                        "description": "Filtra produtos pelo nome",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "descricao",
                        "in": "query",
                        "description": "Filtra produtos pela descrição",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "preco",
                        "in": "query",
                        "description": "Filtra produtos pelo preço exato",
                        "required": false,
                        "schema": {
                            "type": "number",
                            "format": "float"
                        }
                    },
                    {
                        "name": "menorPreco",
                        "in": "query",
                        "description": "Filtra produtos com preço maior ou igual a esse valor",
                        "required": false,
                        "schema": {
                            "type": "number",
                            "format": "float"
                        }
                    },
                    {
                        "name": "maiorPreco",
                        "in": "query",
                        "description": "Filtra produtos com preço menor ou igual a esse valor",
                        "required": false,
                        "schema": {
                            "type": "number",
                            "format": "float"
                        }
                    },
                    {
                        "name": "estoque",
                        "in": "query",
                        "description": "Filtra produtos pelo estoque",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de produtos filtrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer"
                                            },
                                            "nome": {
                                                "type": "string"
                                            },
                                            "descricao": {
                                                "type": "string"
                                            },
                                            "preco": {
                                                "type": "number",
                                                "format": "float"
                                            },
                                            "estoque": {
                                                "type": "integer"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Produtos"
                ],
                "summary": "Cria um novo produto",
                "description": "Cria um produto com os dados fornecidos.",
                "operationId": "createProduto",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nome",
                                    "descricao",
                                    "preco"
                                ],
                                "properties": {
                                    "nome": {
                                        "description": "Nome do produto",
                                        "type": "string",
                                        "example": "Produto X"
                                    },
                                    "descricao": {
                                        "description": "Descrição do produto",
                                        "type": "string",
                                        "example": "Descrição do produto X"
                                    },
                                    "preco": {
                                        "description": "Preço do produto",
                                        "type": "number",
                                        "format": "float",
                                        "example": "19.99"
                                    },
                                    "estoque": {
                                        "description": "Quantidade de estoque do produto",
                                        "type": "integer",
                                        "example": 100
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Produto criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "description": "ID do produto",
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "nome": {
                                            "description": "Nome do produto",
                                            "type": "string",
                                            "example": "Produto X"
                                        },
                                        "descricao": {
                                            "description": "Descrição do produto",
                                            "type": "string",
                                            "example": "Descrição do produto X"
                                        },
                                        "preco": {
                                            "description": "Preço do produto",
                                            "type": "number",
                                            "format": "float",
                                            "example": "19.99"
                                        },
                                        "estoque": {
                                            "description": "Quantidade de estoque do produto",
                                            "type": "integer",
                                            "example": 100
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro na validação dos dados enviados",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Mensagem de erro",
                                            "type": "string",
                                            "example": "Validation failed"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/produtos/{produto}": {
            "get": {
                "tags": [
                    "Produtos"
                ],
                "summary": "Obtém os detalhes de um produto",
                "description": "Retorna as informações do produto especificado pelo ID.",
                "operationId": "getProdutoById",
                "parameters": [
                    {
                        "name": "produto",
                        "in": "path",
                        "description": "ID do produto a ser consultado",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalhes do produto retornados com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "nome": {
                                            "type": "string",
                                            "example": "Produto A"
                                        },
                                        "descricao": {
                                            "type": "string",
                                            "example": "Descrição do Produto A"
                                        },
                                        "preco": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 99.99
                                        },
                                        "estoque": {
                                            "type": "integer",
                                            "example": 50
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Produto não encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Produto não encontrado."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Insira o token",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}